+++ engine/engine-man/src/main/java/com/navinfo/dataservice/engine/man/service/ManApiImpl.java
@@ -452,10 +452,10 @@ public class ManApiImpl implements ManApi {
 	/**
 	 * 查询MAN_TIMELINE
 	 * objName:program,task,subtask,infor
-	 * @return
+	 * @return	Map<Long,Map<String, Object>> key:objId
 	 * @throws ServiceException 
 	 */
-	public Map<Long,Map<String, Object>> queryManTimelineByObjName(String objName) throws Exception{
+	public Map<Integer,Map<String, Object>> queryManTimelineByObjName(String objName) throws Exception{
 		return TimelineService.queryManTimelineByObjName(objName);
 	}
